==> 16S/R_analyses_AD_MS/04_Figure_CDV.Rmd <==
---
title: "Figures CDV"
author: "SRomano"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r setup}
library(phyloseq)
library(tidyverse)
library(ggplot2)
library(SIAMCAT)

set.seed(56987)
```

# Load data
```{r}
loso<-readRDS("/16S/R_analyses_AD/RDS/LOSO_CDV.left.out.rds")
loso$LOSO<-rep("LOSO", nrow(loso))

cdv<-readRDS("/16S/R_analyses_AD/RDS/CDV.rds")
cdv$LOSO<-rep("Single-study models", nrow(cdv))
```

# Formal for plotting
```{r}
df.loso<-data.frame(study = c(as.character(cdv$testing.data), loso$testing.data),
                    percentage_pred = c(cdv$perc.pred, loso$perc.pred),
                    type = c(as.character(cdv$LOSO), loso$LOSO))

df.loso$disease<-vector(length = nrow(loso))
df.loso[which(df.loso$study == "Li" | 
             df.loso$study == "Ling" | 
             df.loso$study == "Zhuang" | 
             df.loso$study == "Yildirim" | 
             df.loso$study == "Vogt" ),]$disease<-"AD"
df.loso[which(df.loso$disease == FALSE),]$disease<-"MS"


# add date to the names
df.loso$study[df.loso$study == "Cekanaviciute"] <- "Cekanaviciute_2017"
df.loso$study[df.loso$study == "Choileain"] <- "Choileain_2020"
df.loso$study[df.loso$study == "Cox"] <- "Cox_2021"
df.loso$study[df.loso$study == "Jangi"] <- "Jiangi_2016"
df.loso$study[df.loso$study == "Forbes"] <- "Forbes_2018"
df.loso$study[df.loso$study == "Vogt"] <- "Vogt_2017"
df.loso$study[df.loso$study == "Ling"] <- "Ling_2021"
df.loso$study[df.loso$study == "Yildirim"] <- "Yildirim_2022"
df.loso$study[df.loso$study == "Li"] <- "Li_2019"
df.loso$study[df.loso$study == "Zhuang"] <- "Zhuang_2018"

df.loso$type<-factor(df.loso$type, levels = c("Single-study models", "LOSO"), ordered = T)
df.loso$percentage_pred<-as.numeric(df.loso$percentage_pred)

# make another df for plotting
(df.loso.average<-df.loso %>%
  group_by(type) %>%
  summarise(mean = mean(percentage_pred)))
   
# make df with average and stdv
(df.loso.summary<-df.loso %>%
  group_by(type, study) %>%
  summarise(average = mean(percentage_pred),
            stedev = sd(percentage_pred)))
```
 
# Do a stats test
```{r}
t.test(percentage_pred ~ type, data = df.loso, var.equal = F)
df.loso %>%
  group_by(type) %>%
  summarise(mean = mean(percentage_pred),
            stdev = sd(percentage_pred))
   
```
# Make a plot with points
```{r}
(pp<-ggplot() + 

   # add lines with averages
   geom_hline(yintercept = subset(df.loso.average, type == "Single-study models")$mean, 
             linetype = "solid", linewidth = 1.3, color = "#FFA258")+
   geom_hline(yintercept = subset(df.loso.average, type == "LOSO")$mean, 
             linetype = "solid", linewidth = 1.3, color = "#CC79A7")  +
   # add points
    
   geom_point(data = df.loso, 
              mapping = aes(x = study, 
                            y = percentage_pred,
                            fill = type),
              pch =21, size = 3.5, alpha = 0.4,
              position = position_dodge2(width = 0.7, 
                                         preserve = "total",
                                         padding = 10), show.legend = F) +
   scale_fill_manual(values = c("#E6AA6E", "#BE82A0")) +

   # add error bar with margin
   geom_errorbar(data = df.loso.summary,
                  mapping = aes(x = study,
                                ymin = average - stedev, ymax = average + stedev, group = type), 
                  width = 0,  color = "black", size = 1.7,
                  position = position_dodge(width = 0.7)) +

   geom_errorbar(data = df.loso.summary,
                  mapping = aes(x = study,
                                ymin = average - stedev, 
                                ymax = average + stedev, 
                                group = type,
                                color = type), 
                  width = 0,  size = 1.45,
                  position = position_dodge(width = 0.7),show.legend = F) +
   # add stats summaries
   geom_point(data = df.loso.summary, 
               mapping = aes(study, 
                           average,
                           fill = type),
               pch = 21,
               size = 6,
               position = position_dodge2(width = 0.7)) +
 
  # change colors
  scale_y_continuous(breaks = seq(0, 100, by = 10), limits =c(-7.5,100)) +
  labs(fill = "ML approach") +
  ggnewscale::new_scale_fill() +
  scale_fill_manual(values = c("#FFA258", "#CC79A7")) +
  scale_color_manual(values = c("#FFA258", "#CC79A7")) +

  # change themes    
  theme_bw() +
  theme(axis.text = element_text(size = 14),
        axis.title = element_text(size =14),
        title = element_text(size =16),
        axis.text.x = element_text(angle = 30, vjust = 1, hjust = 1),
        legend.background = element_blank(),
        legend.text = element_text(size = 12),
        panel.grid = element_blank()) +
  geom_hline(yintercept = 10, 
             linetype = "dashed", linewidth = 1.3, color = "gray") + 
  ggnewscale::new_scale_fill() + # Important to put this after you defined the first scale
  geom_tile(data = df.loso, aes(x =study, 
                                y = -5, 
                                fill = disease, 
                                width = 1, 
                                height = 3), 
            color="black", 
            size = 0.3) +
  scale_fill_manual(values = c("#11793E", "#FFC107")) +
  ylab("False-positive rate") +
  xlab("Studies") +
    labs(fill = "Disease"))

ggsave("/16S/R_analyses_AD/Figures/CDV_points.svg", 
        width = 12, height = 6)


```
